Build successful: Added battle log updating. When running the game, go to button menu and step through actions to see a counter but also witness the past counts decend on the screen.
Implemented obersableList<string> and an extra function in the controller to always add a new action to the front of the list.

Here the engine will be return actions to the list so the log will output all actions being performed such as dungeon start, somebody hit somebody and somebody died.